Move addon into t package and fix imports
