+++ t/operators/__init__.py
@@ -1,4 +1,4 @@
 # Absolute import to the tracking operator package
-from tracking.operators import tracking
+from t.operators import tracking
 
 operator_classes = tracking.operator_classes
