+++ t/operators/tracking/camera.py
