+++ t/ui/__init__.py
@@ -1,4 +1,4 @@
 # Absolute import for the panel aggregator
-from tracking.ui import panels as panels
+from t.ui import panels as panels
 
 panel_classes = panels.panel_classes
