+++ t/ui/panels/__init__.py
@@ -1,5 +1,5 @@
 # Absolute imports for panel modules
-from tracking.ui.panels import tracking_panel, settings_panel, test_panels
+from t.ui.panels import tracking_panel, settings_panel, test_panels
 
 panel_classes = (
     *tracking_panel.panel_classes,
